pyproject.toml
[project]
name = "heygptcli"
version = "0.3.8"
readme = "README.md"
description = "CLI tool to generate text using GPT or Palm based on ready made templated prompts."
authors = [
  { name = "Rishang", email = "rishangbhavsarcs@gmail.com" }
]

requires-python = ">=3.10,<4.0"

dependencies = [
    "requests",
    "fastapi",
    "uvicorn",
    "typer",
    "python-dotenv>=1.0.0",
    "rich",
    "pyarrow",
    "streamlit",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0.3",
    "litellm>=1.80.5",
    "pyperclip>=1.11.0",
    "setuptools>=80.9.0",
]

[dependency-groups]
dev = [
    "pre-commit>=3.3.2",
    "mypy>=1.3.0",
    "types-requests",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"


[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["heygpt*"]

[project.scripts]
heygpt = "heygpt.cli:app"

